app/microservices/market-analysis/requirements.txt
grpc-interceptor >= 0.11.0
grpcio-tools >= 1.30
pymongo >= 3.11.2
dnspython >= 2.1.0
Flask
psycopg2-binary
flask-swagger-ui
pyyaml
prometheus_client
psutil